Fixed crash on items animation drawing.
Other small things.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@193 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 6738c9ac402aec13561108ac8ca4d29c0e1d9418

+++ Source/ItemHex.cpp
@@ -67,7 +67,7 @@ void ItemHex::Process()
 		}
 		else
 		{
-			short cur_spr=begSpr+((endSpr-begSpr+(begSpr<endSpr?1:-1))*anim_proc)/100;
+			int cur_spr=begSpr+((endSpr-begSpr+(begSpr<endSpr?1:-1))*anim_proc)/100;
 			if(curSpr!=cur_spr) SetSpr(cur_spr);
 		}
 	}
